Delete previous regulation before updating the next regulation to avoid constraint
